remove : mail address login
- メールアドレスとパスワードによるログインを削除
- NODE.ENV === 'development'ではメールアドレスログイン可能（ローカル開発用）
