+++ components/ui/login.vue
@@ -1,44 +1,57 @@
 <script setup lang="ts">
-const props = defineProps<{
-    redirect?: string;
-}>();
-
-const emit = defineEmits(['loginSuccess']);
-
 const token = ref();
-const mode_login = ref(true);
+
+const env = process?.env?.NODE_ENV;
+const mailAddressLogin = ref({ email: '', password: '' });
+const handleLogin = async () => {
+    try {
+        await useLogin(
+            mailAddressLogin.value.email,
+            mailAddressLogin.value.password,
+            token.value
+        );
+        navigateTo('/');
+    } catch (error) {
+        console.error(error);
+    }
+};
 </script>
 
 <template>
     <UCard>
-        <AuthLogin
-            v-if="mode_login"
-            v-model="token"
-            :redirect="props.redirect"
-            @success="emit('loginSuccess')"
-        />
+        <div class="min-w-80 flex flex-col gap-4 items-center">
+            <div class="font-bold text-2xl mb-4">ログイン</div>
 
-        <AuthSignup v-else v-model="token" />
+            <ButtonBase
+                :disabled="!token"
+                variant="outline"
+                label="X アカウントでログイン"
+                icon="simple-icons:x"
+                class="w-full"
+                @click="useLoginWithTwitter"
+            />
 
-        <NuxtTurnstile v-model="token" />
-
-        <template #footer>
-            <div class="w-full flex items-center justify-center gap-8">
-                <UiTooltip v-if="mode_login" text="準備中">
-                    <UButton disabled variant="link" :padded="false">
-                        パスワードを忘れた場合
-                    </UButton>
-                </UiTooltip>
-                <UButton
-                    variant="link"
-                    :padded="false"
-                    @click="mode_login = !mode_login"
-                >
-                    {{
-                        mode_login ? 'サインアップはこちら' : 'ログインはこちら'
-                    }}
-                </UButton>
+            <div
+                v-if="env === 'development'"
+                class="w-full flex flex-col gap-2"
+            >
+                <p>Development only</p>
+                <UiTextinput v-model="mailAddressLogin.email" />
+                <UiTextinput
+                    v-model="mailAddressLogin.password"
+                    type="password"
+                />
+                <ButtonBase
+                    :disabled="!token"
+                    variant="outline"
+                    label="Mail address login"
+                    icon="lucide:mail"
+                    class="w-full"
+                    @click="handleLogin"
+                />
             </div>
-        </template>
+
+            <NuxtTurnstile v-model="token" />
+        </div>
     </UCard>
 </template>
